+++ core/src/avm2/op.rs
@@ -32,7 +32,7 @@ pub enum Op<'gc> {
         num_args: u32,
     },
     CallMethod {
-        index: u32,
+        index: usize,
         num_args: u32,
         push_return_value: bool,
     },
@@ -92,7 +92,7 @@ pub enum Op<'gc> {
         num_args: u32,
     },
     ConstructSlot {
-        index: u32,
+        index: usize,
         num_args: u32,
     },
     ConstructSuper {
@@ -174,7 +174,7 @@ pub enum Op<'gc> {
     },
     GetSlot {
         // note: 0-indexed, as opposed to FP.
-        index: u32,
+        index: usize,
     },
     GetSuper {
         multiname: Gc<'gc, Multiname<'gc>>,
@@ -289,7 +289,7 @@ pub enum Op<'gc> {
     RShift,
     SetGlobalSlot {
         // note: 0-indexed, as opposed to FP.
-        index: u32,
+        index: usize,
     },
     SetLocal {
         index: u32,
@@ -307,11 +307,11 @@ pub enum Op<'gc> {
     },
     SetSlot {
         // note: 0-indexed, as opposed to FP.
-        index: u32,
+        index: usize,
     },
     SetSlotNoCoerce {
         // note: 0-indexed, as opposed to FP.
-        index: u32,
+        index: usize,
     },
     SetSuper {
         multiname: Gc<'gc, Multiname<'gc>>,
